+++ include/input/inputFileParser/QMInputParser.hpp
@@ -52,7 +52,7 @@ namespace input
         void parseDispersion(const pq::strings &, const size_t);
         void parseRemoveNetForce(const pq::strings &, const size_t);
 
-        void parseMaceModelSize(const pq::strings &, const size_t);
+        void parseMaceModel(const pq::strings &, const size_t);
         void parseMaceModelPath(const pq::strings &, const size_t);
         void parseMaceQMMethod(const std::string_view &);
 
@@ -62,6 +62,9 @@ namespace input
         void parseHubbardDerivs(const pq::strings &, const size_t);
 
         void parseXtbMethod(const pq::strings &, const size_t);
+
+        void parseFennolModelPath(const pq::strings &, const size_t);
+        void parseGPUPreprocessing(const pq::strings &, const size_t);
     };
 
 }   // namespace input