+++ src/engine/qmmdEngine.cpp
@@ -34,9 +34,10 @@
 #include "turbomoleRunner.hpp"   // for TurbomoleRunner
 
 #ifdef WITH_ASE
-#include "aseDftbRunner.hpp"   // for aseDftbRunner
-#include "aseXtbRunner.hpp"    // for aseXtbRunner
-#include "maceRunner.hpp"      // for MaceRunner
+#include "aseDftbRunner.hpp"     // for aseDftbRunner
+#include "aseFennolRunner.hpp"   // for AseFennolRunner
+#include "aseMaceRunner.hpp"     // for AseMaceRunner
+#include "aseXtbRunner.hpp"      // for aseXtbRunner
 #endif
 
 using engine::QMMDEngine;
@@ -81,6 +82,9 @@ void QMMDEngine::setQMRunner(const QMMethod method)
     else if (method == MACE)
         setMaceQMRunner();
 
+    else if (method == FENNOL)
+        setAseFennolRunner();
+
     else
         throw InputFileException(
             "A qm based jobtype was requested but no external "
@@ -103,16 +107,44 @@ void QMMDEngine::setMaceQMRunner()
     const auto useDFTD   = QMSettings::useDispersionCorr();
     const auto fpType    = Settings::getFloatingPointPybindString();
 
-    auto maceModel = string(QMSettings::getMaceModelSize());
+    auto maceModel = string(QMSettings::getMaceModel());
 
     if (!modelPath.empty())
         maceModel = modelPath;
 
-    _qmRunner = make_shared<MaceRunner>(modelType, maceModel, fpType, useDFTD);
+    _qmRunner =
+        make_shared<AseMaceRunner>(modelType, maceModel, fpType, useDFTD);
 #else
     throw CompileTimeException(
         "A mace type qm method was requested but ASE was not enabled at "
         "compile time. Please recompile with ASE enabled to use mace type "
+        "qm methods using: -DBUILD_WITH_ASE=ON"
+    );
+#endif
+}
+
+/**
+ * @brief sets the QMRunner object for fennol type qm methods.
+ *
+ * @throws InputFileException if ASE was not enabled at compile
+ * time.
+ *
+ */
+void QMMDEngine::setAseFennolRunner()
+{
+#ifdef WITH_ASE
+    using enum FPType;
+
+    const auto modelPath        = QMSettings::getFennolModelPath();
+    const auto gpuPreprocessing = QMSettings::useGPUPreprocessing();
+    const bool useFloat64       = Settings::getFloatingPointType() == DOUBLE;
+
+    _qmRunner =
+        make_shared<AseFennolRunner>(modelPath, gpuPreprocessing, useFloat64);
+#else
+    throw CompileTimeException(
+        "A fennol type qm method was requested but ASE was not enabled at "
+        "compile time. Please recompile with ASE enabled to use fennol type "
         "qm methods using: -DBUILD_WITH_ASE=ON"
     );
 #endif