+++ src/setup/qmSetup.cpp
@@ -75,6 +75,8 @@ QMSetup::QMSetup(QMMDEngine &engine) : _engine(engine) {}
  */
 void QMSetup::setup()
 {
+    setupQMMethodFennol();
+
     setupQMMethodMace();
 
     setupQMMethod();
@@ -122,6 +124,23 @@ void QMSetup::setupQMMethodAseDftbPlus()
 }
 
 /**
+ * @brief setup the FeNNol method of the system
+ *
+ */
+void QMSetup::setupQMMethodFennol()
+{
+    if (QMSettings::getQMMethod() != QMMethod::FENNOL)
+        return;
+
+    if (QMSettings::getFennolModelPath() == "")
+        throw InputFileException(
+            "The FeNNol QM runner has been selected but the "
+            "\"fennol_model_path\" keyword has not been set. This setup is "
+            "invalid."
+        );
+}
+
+/**
  * @brief setup the MACE method of the system
  *
  */
@@ -132,19 +151,20 @@ void QMSetup::setupQMMethodMace()
 
     if (QMSettings::getMaceModelType() != MaceModelType::MACE_MP)
     {
-        const auto modelSize = QMSettings::getMaceModelSize();
-        if (modelSize != MaceModelSize::SMALL &&
-            modelSize != MaceModelSize::MEDIUM &&
-            modelSize != MaceModelSize::LARGE)
-            throw InputFileException(std::format(
-                "The '{}' model size is only compatible with the '{}' "
-                "model type.",
-                string(modelSize),
-                string(MaceModelType::MACE_MP)
-            ));
+        const auto modelSize = QMSettings::getMaceModel();
+        if (modelSize != MaceModel::SMALL && modelSize != MaceModel::MEDIUM &&
+            modelSize != MaceModel::LARGE)
+            throw InputFileException(
+                std::format(
+                    "The '{}' model size is only compatible with the '{}' "
+                    "model type.",
+                    string(modelSize),
+                    string(MaceModelType::MACE_MP)
+                )
+            );
     }
 
-    if (QMSettings::getMaceModelSize() == MaceModelSize::CUSTOM &&
+    if (QMSettings::getMaceModel() == MaceModel::CUSTOM &&
         QMSettings::getMaceModelPath().empty())
         throw InputFileException(
             "You have requested a custom MACE model but haven't provided a "
@@ -152,7 +172,7 @@ void QMSetup::setupQMMethodMace()
             "This setup is invalid."
         );
 
-    if (QMSettings::getMaceModelSize() != MaceModelSize::CUSTOM &&
+    if (QMSettings::getMaceModel() != MaceModel::CUSTOM &&
         !QMSettings::getMaceModelPath().empty())
         throw InputFileException(
             "You have set a custom MACE model path without requesting a custom "
@@ -313,7 +333,7 @@ void QMSetup::setupWriteInfo() const
     if (qmMethod == MACE)
     {
         const auto modelType = QMSettings::getMaceModelType();
-        const auto modelSize = QMSettings::getMaceModelSize();
+        const auto modelSize = QMSettings::getMaceModel();
         const auto modelPath = QMSettings::getMaceModelPath();
         const auto fp        = Settings::getFloatingPointPybindString();
         const auto useDisp   = QMSettings::useDispersionCorr() ? "on" : "off";
@@ -329,11 +349,30 @@ void QMSetup::setupWriteInfo() const
         logOutput.writeSetupInfo(modelTypeMsg);
         logOutput.writeSetupInfo(modelSizeMsg);
 
-        if (modelSize == MaceModelSize::CUSTOM)
+        if (modelSize == MaceModel::CUSTOM)
             logOutput.writeSetupInfo(modelPathMsg);
 
         logOutput.writeSetupInfo(fpMsg);
         logOutput.writeSetupInfo(dispCorrMsg);
+    }
+
+    if (qmMethod == FENNOL)
+    {
+        using enum FPType;
+
+        const auto modelPath           = QMSettings::getFennolModelPath();
+        const auto useGPUPreprocessing = QMSettings::useGPUPreprocessing();
+        const bool useFloat64 = Settings::getFloatingPointType() == DOUBLE;
+
+        // clang-format off
+        const auto modelPathMsg  = std::format("Model path:               {}", modelPath);
+        const auto gpuPreprocMsg = std::format("Using GPU pre-processing: {}", useGPUPreprocessing);
+        const auto fpMsg         = std::format("Using float64:            {}", useFloat64);
+        // clang-format on
+
+        logOutput.writeSetupInfo(modelPathMsg);
+        logOutput.writeSetupInfo(gpuPreprocMsg);
+        logOutput.writeSetupInfo(fpMsg);
     }
 
     if (qmMethod == ASEDFTBPLUS)