+++ src/setup/setup.cpp
@@ -225,4 +225,6 @@ void setup::setupEngine(Engine &engine)
 
     if (Settings::isOptJobType())
         setupOptimizer(engine);
+
+    engine.getLogOutput().flushQueuedWarnings();
 }